Deepen STS transformer architecture; bump sts to 0.6.11
Architecture pass on the STS transformer (built on top of the NISO STS
v1.2 migration in 279f8e64):

- Split HtmlRenderer::Ruby god-class: extract per-node Handlers module
  (html_renderer/handlers.rb, 526 LOC) from the 900-LOC ruby.rb. The
  Ruby class now holds only the page-shell, liquid, and walker glue.

- Commit to ModelBuilder as the single construction site. Replace
  ~70 direct Sts::NisoSts::* / Sts::TbxIsoTml::* construction sites
  across inline/term/iso_meta/back/front/body/section/footnote transformers
  with ModelBuilder factory calls, and switch the factories themselves
  from Klass.new.tap { } to Klass.new do |x| ... end so they go through
  lutaml-model builder-tracker (which records element_order correctly
  for direct setters as of lutaml-model 0.8.19).

- Collapse duplicate inline dispatchers in InlineTransformer. The
  transform_inline / transform_tbx_inline case/when trees shared most
  of their branches; merge them through a shared apply_inline_content
  path keyed on a single TEXT_ATTR_FOR table.

- Stop testing private transformer methods. Rewrite the STS specs to
  drive the public API (Metanorma::Iso::Sts.convert) instead of poking
  at transformer.send(:transform_X, ...). The interface is the test
  surface.

- Bump sts dependency to ~> 0.6.11. sts 0.6.11 declares the repeatable
  children of Sts::NisoSts::MetadataIso (std_ref, title_wrap, ics,
  release_date, permissions, meta_date, std_xref, content_language,
  custom_meta_group) with collection: true. Earlier sts versions
  declared them as singular attributes, so the appender form
  silently dropped all but the last value; the recent lutaml-model
  builder-tracker fix (0.8.19) surfaced this as duplicate-content
  serialization.
